full_stream/src/trainer/requirements.txt
kafka_python==2.0.2
python-dotenv==0.15.0
numpy==1.19.2
scipy==1.5.2
